feat(user): optimize repository counts and use BigDecimal for earnings
- Added explicit JPQL queries for countAll and countByRole in AppUserRepository to guarantee efficient performance.
- Updated AppUserStats entity and StatsDTO to use BigDecimal for totalEarnings to ensure financial accuracy.
- Confirmed full support for Admin module's dashboard requirements.

+++ backend/src/main/java/com/deliverXY/backend/NewCode/user/dto/StatsDTO.java
@@ -2,9 +2,11 @@ package com.deliverXY.backend.NewCode.user.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StatsDTO {
     private Double rating;
     private Integer totalDeliveries;
-    private Double totalEarnings;
+    private BigDecimal totalEarnings;
 }
